refactor(cmd/osty): clean up split fallout — relocate runResolveFile, drop orphan docs, revert binary
The previous split moved bodies but left:
- ~140 lines of orphan doc comments stranded in main.go where the
  function bodies had moved to native_check.go / lint_cmd.go
- runResolveFile + nativeLazyStdlibProvider.LookupPackage left in
  main.go even though they're part of the native check pipeline
- runResolveFile's doc accidentally glued to runTypecheckPackageNative
  in native_check.go (Go reads two consecutive comment blocks as one
  doc — godoc would show the wrong description)
- the osty bootstrap binary swapped to a 26MB rebuild and the
  per-worktree .claude metadata file checked in

Move runResolveFile + LookupPackage to native_check.go alongside the
other native-pipeline helpers, delete the orphaned comments, restore
the binary to the main-tree blob, and gitignore .claude/worktrees/
so future worktree scratch never gets committed.

main.go: -221, native_check.go: +75 net (mostly the moved function).
Pure refactor — build, vet, lexer/parser/resolve/check/diag/format/
lint/pipeline test packages, and the cmd/osty refactor-touched tests
all still pass.
